Implemented code review
