+++ src/test/java/testresults/CVSB_981/AutoCalculateAbandonedTestResults_CVSB_2208.java
@@ -6,12 +6,13 @@ import net.thucydides.core.annotations.Title;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import pages.SelectReasonPage;
+import pages.TestPage;
 import steps.*;
 import steps.composed.TestTypeCategoryComp;
 import util.BaseTestClass;
 
 @RunWith(SerenityRunner.class)
-public class AutoCalculateAbandonedTestResults_2209 extends BaseTestClass {
+public class AutoCalculateAbandonedTestResults_CVSB_2208 extends BaseTestClass {
 
     @Steps
     TestSteps testSteps;
@@ -23,39 +24,24 @@ public class AutoCalculateAbandonedTestResults_2209 extends BaseTestClass {
     TestTypeCategoryComp testTypeCategoryComp;
 
     @Steps
-    EUVehicleCategorySteps euVehicleCategorySteps;
-
-    @Steps
-    OdometerReadingSteps odometerReadingSteps;
-
-    @Steps
-    TestReviewSteps testReviewSteps;
-
-    @Steps
     AbandonTestSteps abandonTestSteps;
 
     @Steps
     SelectReasonSteps selectReasonSteps;
 
-    @Title("CVSB_981 - AC1 Auto-calculate test results upon abandoning a test type (test review screen)")
+
+    @Title("CVSB_981 - AC1 Auto-calculate test results upon abandoning a test type (test overview screen)")
     @Test
-    public void testAutoCalculateAbandonedTestInTestReview() {
+    public void testAbandonTestTypeInTestOverview() {
         testTypeCategoryComp.goToTestPage("TC7524","Avello Edinburgh Ltd");
-        testSteps.selectVehicleCategoryOption();
-        euVehicleCategorySteps.selectM1Option();
-        testSteps.selectOdometerReading();
-        odometerReadingSteps.typeInField("5");
-        odometerReadingSteps.pressSave();
         testSteps.addTestType();
         testTypeCategorySteps.selectFromTestTypeList("Annual test");
         testSteps.swipeTestType("Annual test");
         testSteps.pressTestTypeAbandonButton();
-        selectReasonSteps.selectMultipleReasons(SelectReasonPage.Reasons.REASON_8, SelectReasonPage.Reasons.REASON_12,
-                SelectReasonPage.Reasons.REASON_6, SelectReasonPage.Reasons.REASON_2);
+        selectReasonSteps.selectAReason(SelectReasonPage.Reasons.REASON_1);
         selectReasonSteps.pressNextButton();
         abandonTestSteps.pressDone();
         abandonTestSteps.pressAbandon();
-        testSteps.reviewAction();
-        testReviewSteps.checkTestStatus("Annual test", "ABANDONED");
+        testSteps.checkTestTypeStatus("Annual test", TestPage.TestTypeStatuses.ABANDONED);
     }
 }
