+++ src/test/java/testresults/CVSB_981/AutoCalculateAbandonedTestResults_CVSB_2212.java
@@ -12,7 +12,7 @@ import steps.composed.TestTypeCategoryComp;
 import util.BaseTestClass;
 
 @RunWith(SerenityRunner.class)
-public class AutoCalculateAbandonedTestResults_2212 extends BaseTestClass {
+public class AutoCalculateAbandonedTestResults_CVSB_2212 extends BaseTestClass {
 
     @Steps
     TestSteps testSteps;
